Cleaned up rebasing and refactoring.
