+++ src/test/scala/org/lamastex/spark/trendcalculus/ScalableTest.scala
@@ -1,4 +1,4 @@
-package org.lamastex.trendcalculus
+package org.lamastex.spark.trendcalculus
 
 import org.scalatest.Matchers
 
@@ -12,11 +12,9 @@ class ScalableTest extends SparkSpec with Matchers {
     import org.apache.spark.sql.functions._
     import spark.implicits._
 
-    import org.lamastex.trendcalculus.DateUtils.Frequency
-    import org.lamastex.trendcalculus.SeriesUtils.FillingStrategy
-    import org.lamastex.trendcalculus._
+    import org.lamastex.spark.trendcalculus._
 
-    val filePathRoot: String = "file:///root/GIT/lamastex/spark-trend-calculus/src/test/resources/org/lamastex/trendcalculus/"
+    val filePathRoot: String = "src/test/resources/org/lamastex/spark/trendcalculus/"
     val df = spark.read.option("header", "true").option("inferSchema", "true").csv(filePathRoot+"brent.csv").filter(year(col("DATE")) >= 2015)
 
     val toMilliUDF = udf( { time: java.sql.Timestamp => time.getTime() } )
